WIP: Extract `ui` crate
