--- crates/storybook/src/ui.rs
@@ -1,7 +1,0 @@
-mod component;
-mod element;
-mod module;
-
-pub use component::*;
-pub use element::*;
-pub use module::*;
