+++ crates/ui/src/elements/icon.rs
@@ -1,8 +1,9 @@
-use crate::theme::theme;
 use gpui2::elements::svg;
 use gpui2::style::StyleHelpers;
 use gpui2::IntoElement;
 use gpui2::{Element, ViewContext};
+
+use crate::theme;
 
 // Icon::Hash
 // icon(IconAsset::Hash).color(IconColor::Warning)
