+++ crates/ui/src/modules/chat_panel.rs
