+++ crates/ui/src/modules/project_panel.rs
