+++ crates/ui/src/modules/status_bar.rs
