+++ crates/ui/src/modules/tab_bar.rs
